[red-knot] Unify setup_db functions
