+++ crates/red_knot_python_semantic/src/lib.rs
@@ -25,4 +25,7 @@ pub mod types;
 mod unpack;
 mod util;
 
+#[cfg(test)]
+mod test;
+
 type FxOrderSet<V> = ordermap::set::OrderSet<V, BuildHasherDefault<FxHasher>>;
