+++ crates/red_knot_python_semantic/src/symbol.rs
@@ -90,7 +90,7 @@ impl<'db> Symbol<'db> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::types::tests::setup_db;
+    use crate::test::setup_db;
 
     #[test]
     fn test_symbol_or_fall_back_to() {
